requirements.txt
asgiref~=3.8.1
Django~=5.1.1
django-cors-headers~=4.4.0
djangorestframework~=3.15.2
djangorestframework-simplejwt~=5.3.1
PyJWT~=2.9.0
pytz~=2024.1
sqlparse~=0.5.1
psycopg2-binary~=2.9.9
python-dotenv~=1.0.1
pillow~=10.4.0
pip~=24.2
docutils~=0.21.2
filelock~=3.15.4
setuptools~=72.1.0
botocore~=1.35.32
boto3~=1.35.32
s3transfer~=0.10.2
jmespath~=1.0.1
requests~=2.32.3
pywin32~=306
urllib3~=2.2.2
python-dateutil~=2.9.0.post0
certifi~=2024.7.4
channels~=4.1.0
six~=1.16.0
django-storages~=1.14.4
idna~=3.7
platformdirs~=4.2.2
Pygments~=2.18.0